solution to eyden-tracer-04
I couldn't implement CSolidGeosphere class completely, but I tried to adjust the code from the link I mentioned in comments. 

For Phong Shader with bump mapping procedure, I just adjusted CShaderPhong, but did not change the name. Also, I tried to use ground.obj to test it, but unfortunately, was not successful.